Refactored a lot of the pytorch graph code and simplified the logging
