+++ wandb/data_types.py
@@ -9,34 +9,6 @@ import six
 import wandb
 import numpy as np
 from wandb import util
-from operator import mul
-from six.moves import reduce
-
-
-def nested_shape(array_or_tuple):
-    """Figures out the shape of tensors possibly embedded in tuples
-     i.e 
-     [0,0] returns (2)
-     ([0,0], [0,0]) returns (2,2)
-     (([0,0], [0,0]),[0,0]) returns ((2,2),2)
-     """
-    if hasattr(array_or_tuple, 'size'):
-        # pytorch tensors use V.size() to get size of tensor
-        return list(array_or_tuple.size())
-    elif hasattr(array_or_tuple, 'get_shape'):
-        # tensorflow uses V.get_shape() to get size of tensor
-        return array_or_tuple.get_shape().as_list()
-    elif hasattr(array_or_tuple, 'shape'):
-        return array_or_tuple.shape
-
-    try:
-        #treat object as iterable
-        return [nested_shape(item) for item in list(array_or_tuple)]
-    except TypeError:
-        # object is not actually iterable
-        # LB: Maybe we should throw an error?
-        return []
-
 
 def val_to_json(key, val, mode="summary", step=None):
     """Converts a wandb datatype to its JSON representation"""
@@ -202,219 +174,6 @@ class Graph(object):
 
         return graph
 
-    @classmethod
-    def hook_torch(cls, model, criterion=None):
-        graph = cls("torch")
-        graph.hook_torch_modules(model, criterion)
-        return graph
-
-    def create_forward_hook(self, name, modules):
-        graph = self
-
-        def after_forward_hook(module, inp, output):
-            if id(module) in modules:
-                return
-            modules.add(id(module))
-            if not isinstance(output, tuple):
-                output = (output,)
-            parameters = [(pname, list(param.size()))
-                          for pname, param in module.named_parameters()]
-
-            node = Node(
-                id=id(module),
-                name=name,
-                class_name=str(module),
-                output_shape=nested_shape(output),
-                parameters=parameters,
-                num_parameters=[reduce(mul, size)
-                                for (pname, size) in parameters]
-            )
-            graph.nodes_by_id[id(module)] = node
-            for param in module.parameters():
-                graph.nodes_by_id[id(param)] = node
-            graph.add_node(node)
-            if not graph.criterion_passed:
-                graph.criterion = output[0].grad_fn
-
-        return after_forward_hook
-
-    def hook_torch_modules(self, module, criterion=None, prefix=None):
-        torch = util.get_module("torch", "Could not import torch")
-        hooks = []
-        modules = set()
-        layers = 0
-        graph = self
-        if criterion:
-            graph.criterion = criterion
-            graph.criterion_passed = True
-
-        for name, sub_module in module.named_children():
-            name = name or str(layers)
-            if prefix:
-                name = prefix + "." + name
-            layers += 1
-            if not isinstance(sub_module, torch.nn.Module):
-                # TODO: Why does this happen?
-                break
-
-            # Trying to support torch >0.3 making this code complicated
-            # We want a list of types that we should recurse into
-            # Torch 0.3   uses containers
-            #       0.4   has ModuleList
-            #       0.4.1 has ModuleDict
-            module_types = [getattr(torch.nn, module_classname)
-                for module_classname in ("Container", "Sequential", "ModuleList", "ModuleDict")
-                if hasattr(torch.nn, module_classname)]
-                   
-            if isinstance(sub_module, tuple(module_types)):
-                self.hook_torch_modules(sub_module, prefix=name)
-            else:
-                def backward_hook(module, input, output):
-                    [hook.remove() for hook in hooks]
-                    
-                    graph.loaded = True
-                    # TODO: Keeping this here as a starting point for adding graph data
-                    if not graph.loaded:
-                        def traverse(node, functions=[]):
-                            if hasattr(node, 'grad_fn'):
-                                node = node.grad_fn
-
-                            if hasattr(node, 'variable'):
-                                node = graph.nodes_by_id.get(id(node.variable))
-                                if node:
-                                    node.functions = list(functions)
-                                    del functions[:]
-
-                            if hasattr(node, 'next_functions'):
-                                functions.append(type(node).__name__)
-                                for f in node.next_functions:
-                                    if f[0]:
-                                        functions.append(type(f[0]).__name__)
-                                        traverse(f[0], functions)
-
-                            if hasattr(node, 'saved_tensors'):
-                                for t in node.saved_tensors:
-                                    traverse(t)
-
-                        traverse(graph.criterion)
-
-                hooks.append(
-                    sub_module.register_forward_hook(self.create_forward_hook(name, modules)))
-                hooks.append(
-                    sub_module.register_backward_hook(backward_hook))
-
-    @classmethod
-    def from_torch_layers(cls, module_graph, variable):
-        """Recover something like neural net layers from PyTorch Module's and the
-        compute graph from a Variable.
-
-        Example output for a multi-layer RNN. We confusingly assign shared embedding values
-        to the encoder, but ordered next to the decoder.
-
-        rnns.0.linear.module.weight_raw rnns.0
-        rnns.0.linear.module.bias rnns.0
-        rnns.1.linear.module.weight_raw rnns.1
-        rnns.1.linear.module.bias rnns.1
-        rnns.2.linear.module.weight_raw rnns.2
-        rnns.2.linear.module.bias rnns.2
-        rnns.3.linear.module.weight_raw rnns.3
-        rnns.3.linear.module.bias rnns.3
-        decoder.weight encoder
-        decoder.bias decoder
-        """
-        # TODO: We're currently not using this, but I left it here incase we want to resurrect! - CVP
-        torch = util.get_module("torch", "Could not import torch")
-
-        module_nodes_by_hash = {id(n): n for n in module_graph.nodes}
-        module_parameter_nodes = [
-            n for n in module_graph.nodes if isinstance(n.obj, torch.nn.Parameter)]
-
-        names_by_pid = {id(n.obj): n.name for n in module_parameter_nodes}
-
-        reachable_param_nodes = module_graph[0].reachable_descendents()
-        reachable_params = {}
-        module_reachable_params = {}
-        names = {}
-        for pid, reachable_nodes in reachable_param_nodes.items():
-            node = module_nodes_by_hash[pid]
-            if not isinstance(node.obj, torch.nn.Module):
-                continue
-            module = node.obj
-            reachable_params = {}  # by object id
-            module_reachable_params[id(module)] = reachable_params
-            names[node.name] = set()
-            for reachable_hash in reachable_nodes:
-                reachable = module_nodes_by_hash[reachable_hash]
-                if isinstance(reachable.obj, torch.nn.Parameter):
-                    param = reachable.obj
-                    reachable_params[id(param)] = param
-                    names[node.name].add(names_by_pid[id(param)])
-
-        # we look for correspondences between sets of parameters used in subtrees of the
-        # computation graph and sets of parameters contained in subtrees of the module
-        # graph
-        node_depths = {id(n): d for n, d in module_graph[0].descendent_bfs()}
-        parameter_module_names = {}
-        parameter_modules = {}
-        for param_node in (n for n in module_graph.nodes if isinstance(n.obj, torch.nn.Parameter)):
-            pid = id(param_node.obj)
-            best_node = None
-            best_depth = None
-            best_reachable_params = None
-            for node in module_graph.nodes:
-                if not isinstance(node.obj, torch.nn.Module):
-                    continue
-                module = node.obj
-                reachable_params = module_reachable_params[id(module)]
-                if pid in reachable_params:
-                    depth = node_depths[id(node)]
-                    if best_node is None or (len(reachable_params), depth) <= (len(best_reachable_params), best_depth):
-                        #print(param_node.name, node.name)
-                        best_node = node
-                        best_depth = depth
-                        best_reachable_params = reachable_params
-
-            parameter_modules[pid] = best_node
-            parameter_module_names[param_node.name] = best_node.name
-
-        # contains all parameters but only a minimal set of modules necessary
-        # to contain them (and which ideally correspond to conceptual layers)
-        reduced_module_graph = cls()
-        rmg_ids = itertools.count()
-        rmg_root = Node(id=next(rmg_ids), node=module_graph[0])
-        reduced_module_graph.add_node(rmg_root)
-        reduced_module_graph.root = rmg_root
-        rmg_nodes_by_pid = {}
-
-        module_nodes_by_pid = {id(n.obj): n for n in module_graph.nodes}
-
-        compute_graph, compute_node_vars = cls.from_torch_compute_graph(
-            variable)
-        for node, _ in reversed(list(compute_graph[0].ancestor_bfs())):
-            param = compute_node_vars.get(node.id)
-            pid = id(param)
-            if not isinstance(param, torch.nn.Parameter):
-                continue
-            if pid not in module_nodes_by_pid:
-                # not all Parameters that occur in the compute graph come from the Module graph
-                continue
-
-            # add the nodes in the order we want to display them on the frontend
-            mid = id(parameter_modules[pid].obj)
-            if mid in rmg_nodes_by_pid:
-                rmg_module = rmg_nodes_by_pid[mid]
-            else:
-                rmg_module = rmg_nodes_by_pid[mid] = Node(
-                    id=next(rmg_ids), node=module_nodes_by_pid[mid])
-                reduced_module_graph.add_node(rmg_module)
-                reduced_module_graph.add_edge(rmg_root, rmg_module)
-
-            rmg_param = Node(id=next(rmg_ids), node=module_nodes_by_pid[pid])
-            rmg_nodes_by_pid[pid] = rmg_param
-            reduced_module_graph.add_node(rmg_param)
-
-            reduced_module_graph.add_edge(rmg_module, rmg_param)
-        return reduced_module_graph
 
     @staticmethod
     def transform(graph):
@@ -579,19 +338,6 @@ class Node(object):
         node.output_shape = output_shape
         node.num_parameters = layer.count_params()
 
-
-        return node
-
-    @classmethod
-    def from_torch_module(cls, nid, module):
-        numpy = util.get_module("numpy", "Could not import numpy")
-
-        node = cls()
-        node.id = nid
-        node.child_parameters = 0
-        for parameter in module.parameters():
-            node.child_parameters += numpy.prod(parameter.size())
-        node.class_name = type(module).__name__
 
         return node
